random trigger streaming from xrootd?!...no random access

+++ MakeMC.csh
@@ -112,6 +112,8 @@ shift
 setenv eBEAM_CURRENT $1
 shift
 setenv EXPERIMENT $1
+shift
+setenv RANDOM_TRIG_NUM_EVT $1
 
 setenv USER_BC `which bc`
 setenv USER_PYTHON `which python`
@@ -122,7 +124,11 @@ setenv USER_BC '/usr/bin/bc'
 setenv USER_STAT '/usr/bin/stat'
 endif
 
-
+setenv MAKE_MC_USING_XROOTD 0
+if ( -f /usr/lib64/libXrdPosixPreload.so ) then
+	setenv MAKE_MC_USING_XROOTD 1
+	setenv LD_PRELOAD /usr/lib64/libXrdPosixPreload.so
+endif
 
 #necessary to run swif, uses local directory if swif=0 is used
 if ( "$BATCHRUN" != "0"  ) then
@@ -359,9 +365,11 @@ echo ""
 echo "=======SOFTWARE USED======="
 echo "MCwrapper version v"$MCWRAPPER_VERSION
 echo "MCwrapper location" $MCWRAPPER_CENTRAL
+echo "Streaming via xrootd? "$MAKE_MC_USING_XROOTD
 echo "BC "$USER_BC
 echo "python "$USER_PYTHON
 echo `which $GENERATOR`
+
 if ( "$GEANTVER" == "3" ) then
 	echo `which hdgeant`
 else
@@ -469,17 +477,7 @@ if ( "$BKGFOLDSTR" == "DEFAULT" || "$bkgloc_pre" == "loc:" || "$BKGFOLDSTR" == "
 	endif
 
     echo "Finding the right file to fold in during MCsmear step"
-    set runperiod="RunPeriod-2018-01"
-
-    if ( $RUN_NUMBER >= 40000 ) then
-		set runperiod="RunPeriod-2018-01"
-	else if ( $RUN_NUMBER >= 30000 ) then
-		set runperiod="RunPeriod-2017-01"
-	else if ( $RUN_NUMBER >= 20000 ) then
-		set runperiod="RunPeriod-2016-10"
-	else if ( $RUN_NUMBER >= 10000 ) then
-		set runperiod="RunPeriod-2016-02"
-    endif
+ 
 
     if ( $RUN_NUMBER < 10000 ) then
 		echo "Warning: random triggers do not exist for this run"
@@ -489,7 +487,7 @@ if ( "$BKGFOLDSTR" == "DEFAULT" || "$bkgloc_pre" == "loc:" || "$BKGFOLDSTR" == "
 	if ( "$bkgloc_pre" == "loc:" ) then
 		set rand_bkg_loc=`echo $BKGFOLDSTR | cut -c 5-`
 		 if ( "$BATCHSYS" == "OSG" && $BATCHRUN != 0 ) then
-                        set     bkglocstring="/srv/run$formatted_runNumber""_random.hddm"
+                        set bkglocstring="/srv/run$formatted_runNumber""_random.hddm"
 		 else
 		    set bkglocstring=$rand_bkg_loc"/run$formatted_runNumber""_random.hddm"
 		 endif
@@ -502,7 +500,7 @@ if ( "$BKGFOLDSTR" == "DEFAULT" || "$bkgloc_pre" == "loc:" || "$BKGFOLDSTR" == "
 		endif
 	endif
 	
-    if ( ! -f $bkglocstring ) then
+    if ( ! -f $bkglocstring && $MAKE_MC_USING_XROOTD == 0 ) then
 		echo "something went wrong with initialization"
 		echo "Could not find mix-in file "$bkglocstring
 		exit 1
@@ -1073,23 +1071,37 @@ if ( "$GENR" != "0" ) then
 			mcsmear $MCSMEAR_Flags -PTHREAD_TIMEOUT=500 -o$STANDARD_NAME'_geant'$GEANTVER'_smeared.hddm' $STANDARD_NAME'_geant'$GEANTVER'.hddm'
 			set mcsmear_return_code=$status
 	    else if ( "$BKGFOLDSTR" == "DEFAULT" || "$BKGFOLDSTR" == "Random" ) then
-			rm -f count.py
-	    	echo "import hddm_s" > count.py
-	    	echo "print(sum(1 for r in hddm_s.istream('$bkglocstring')))" >>! count.py
-	    	set totalnum=`$USER_PYTHON count.py`
-	    	rm count.py
+			
+			if ( $RANDOM_TRIG_NUM_EVT == -1 ) then
+				rm -f count.py
+	    		echo "import hddm_s" > count.py
+	    		echo "print(sum(1 for r in hddm_s.istream('$bkglocstring')))" >>! count.py
+	    		set totalnum=`$USER_PYTHON count.py`
+	    		rm count.py
+			else
+				set totalnum=$RANDOM_TRIG_NUM_EVT
+			endif
 			set fold_skip_num=`echo "($FILE_NUMBER * $PER_FILE)%$totalnum" | $USER_BC`
 			#set bkglocstring="/w/halld-scifs17exp/halld2/home/tbritton/MCwrapper_Development/converted.hddm"
-			
-			echo "mcsmear "$MCSMEAR_Flags" -PTHREAD_TIMEOUT=500 -o$STANDARD_NAME"\_"geant$GEANTVER"\_"smeared.hddm $STANDARD_NAME"\_"geant$GEANTVER.hddm $bkglocstring"\:"1""+"$fold_skip_num
+			if ( $MAKE_MC_USING_XROOTD == 0 ) then
+				echo "mcsmear "$MCSMEAR_Flags" -PTHREAD_TIMEOUT=500 -o$STANDARD_NAME"\_"geant$GEANTVER"\_"smeared.hddm $STANDARD_NAME"\_"geant$GEANTVER.hddm $bkglocstring"\:"1""+"$fold_skip_num
 				mcsmear $MCSMEAR_Flags -PTHREAD_TIMEOUT=500 -o$STANDARD_NAME\_geant$GEANTVER\_smeared.hddm $STANDARD_NAME\_geant$GEANTVER.hddm $bkglocstring\:1\+$fold_skip_num
+			else
+				echo "mcsmear $MCSMEAR_Flags -PTHREAD_TIMEOUT=500 -o$STANDARD_NAME\_geant$GEANTVER\_smeared.hddm $STANDARD_NAME\_geant$GEANTVER.hddm xroot://scosg16.jlab.org//osgpool/halld/random_triggers/$RANDBGTAG/run$formatted_runNumber\_random.hddm:1\:1\+$fold_skip_num"
+				mcsmear $MCSMEAR_Flags -PTHREAD_TIMEOUT=500 -o$STANDARD_NAME\_geant$GEANTVER\_smeared.hddm $STANDARD_NAME\_geant$GEANTVER.hddm xroot://scosg16.jlab.org//osgpool/halld/random_triggers/$RANDBGTAG/run$formatted_runNumber\_random.hddm:1\:1\+$fold_skip_num
+				
+			endif
 			set mcsmear_return_code=$status
 		else if ( "$bkgloc_pre" == "loc:" ) then
-			rm -f count.py
-	    	echo "import hddm_s" > count.py
-	    	echo "print(sum(1 for r in hddm_s.istream('$bkglocstring')))" >>! count.py
-	    	set totalnum=`$USER_PYTHON count.py`
-	    	rm count.py
+			if ( $RANDOM_TRIG_NUM_EVT == -1 ) then
+				rm -f count.py
+	    		echo "import hddm_s" > count.py
+	    		echo "print(sum(1 for r in hddm_s.istream('$bkglocstring')))" >>! count.py
+	    		set totalnum=`$USER_PYTHON count.py`
+	    		rm count.py
+			else
+				set totalnum=$RANDOM_TRIG_NUM_EVT
+			endif
 			set fold_skip_num=`echo "($FILE_NUMBER * $PER_FILE)%$totalnum" | $USER_BC`
 			echo "mcsmear "$MCSMEAR_Flags" -PTHREAD_TIMEOUT=500 -o$STANDARD_NAME"\_"geant$GEANTVER"\_"smeared.hddm $STANDARD_NAME"\_"geant$GEANTVER.hddm $bkglocstring"\:"1""+"$fold_skip_num
 			mcsmear $MCSMEAR_Flags -PTHREAD_TIMEOUT=500 -o$STANDARD_NAME\_geant$GEANTVER\_smeared.hddm $STANDARD_NAME\_geant$GEANTVER.hddm $bkglocstring\:1\+$fold_skip_num
